feat(core): add validator for stage/trigger providing repository info
