+++ app/scripts/modules/core/src/pipeline/config/validation/repositoryInformationProvided.validator.ts
@@ -1,0 +1,3 @@
+import { buildUpstreamHasFlagValidator } from 'core/pipeline/config/validation/upstreamHasFlagValidator.builder';
+
+buildUpstreamHasFlagValidator('providesRepositoryInformation', 'repositoryInformationProvided');
